% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_stability.R
\name{BatchStabilityEst}
\alias{BatchStabilityEst}
\title{Estimate gene-level batch stability scores}
\usage{
BatchStabilityEst(
  seurat.obj,
  batch = "batch",
  ref.batch = NULL,
  use.prev.ref = TRUE,
  genes = NULL,
  ngenes = NULL,
  K = 5,
  max_iter = 10,
  n.cores = 10,
  seed = 42,
  subsample = NULL
)
}
\description{
Estimate gene-level batch stability scores
}
