% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_stability.R
\name{BatchStableGenes}
\alias{BatchStableGenes}
\title{Select batch-stable genes using an elbow rule}
\usage{
BatchStableGenes(obj, plot = FALSE, include.na = T)
}
\description{
Select batch-stable genes using an elbow rule
}
