% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_methods.R
\name{FastMNN}
\alias{FastMNN}
\title{FastMNN integration wrapper}
\usage{
FastMNN(seurat.obj, new.reduction, ndims, default.assay, verbose = FALSE, ...)
}
\description{
FastMNN integration wrapper
}
