% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_methods.R
\name{Integration}
\alias{Integration}
\title{Run an integration backend on a Seurat object}
\usage{
Integration(
  seurat.obj,
  batch = "batch",
  method,
  new.reduction = NULL,
  ndims = 30,
  verbose = FALSE,
  default.assay = "RNA",
  ...
)
}
\description{
Run an integration backend on a Seurat object
}
