% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_score.R
\name{IntegrationScore.Joint}
\alias{IntegrationScore.Joint}
\title{Score an integrated embedding using a joint multivariate GMM}
\usage{
IntegrationScore.Joint(
  seurat.obj,
  reduction,
  ref.batch,
  ndims = NULL,
  batch = "batch",
  K = 20,
  seed = 42,
  n.cores = 1
)
}
\description{
Score an integrated embedding using a joint multivariate GMM
}
