% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_score.R
\name{IntegrationScore}
\alias{IntegrationScore}
\title{Score an integrated embedding dimension by dimension}
\usage{
IntegrationScore(
  seurat.obj,
  reduction,
  ref.batch,
  ndims = NULL,
  batch = "batch",
  K = 5,
  n.cores = 10,
  seed = 42
)
}
\description{
Score an integrated embedding dimension by dimension
}
