% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Preprocess.LIGER}
\alias{Preprocess.LIGER}
\title{Preprocess a Seurat object for LIGER integration}
\usage{
Preprocess.LIGER(
  seurat.obj,
  batch = "batch",
  genes = NULL,
  ngenes = 2000,
  default.assay = "RNA",
  verbose = FALSE
)
}
\description{
Preprocess a Seurat object for LIGER integration
}
