% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess a Seurat object for PCA-based integration methods}
\usage{
Preprocess(
  seurat.obj,
  batch = "batch",
  ndims = 30,
  genes = NULL,
  ngenes = 2000,
  default.assay = "RNA",
  verbose = FALSE
)
}
\description{
Preprocess a Seurat object for PCA-based integration methods
}
