% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Preprocess.scVI}
\alias{Preprocess.scVI}
\title{Preprocess a Seurat object for scVI integration}
\usage{
Preprocess.scVI(
  seurat.obj,
  batch = "batch",
  genes = NULL,
  ngenes = 2000,
  default.assay = "RNA",
  verbose = FALSE
)
}
\description{
Preprocess a Seurat object for scVI integration
}
