% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_methods.R
\name{Harmony}
\alias{Harmony}
\title{Harmony integration wrapper}
\usage{
Harmony(seurat.obj, new.reduction, ndims, default.assay, verbose = FALSE, ...)
}
\description{
Harmony integration wrapper
}
