% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_search.R
\name{IntegrateRigor.ParameterS}
\alias{IntegrateRigor.ParameterS}
\title{Search integration parameters using marginal integration scores}
\usage{
IntegrateRigor.ParameterS(
  seurat.obj,
  parameter.df,
  method,
  batch = "batch",
  ndims = 30,
  ndims.score = min(30, ndims),
  ref.batch = NULL,
  use.prev.ref = TRUE,
  force.run = TRUE,
  K = 5,
  n.cores = 5,
  seed = 42,
  return.all = TRUE,
  use.batch.stable.genes = TRUE,
  verbose = FALSE,
  subsample = NULL,
  run.optimal = TRUE,
  default.assay = "RNA"
)
}
\description{
Search integration parameters using marginal integration scores
}
