% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.parameter.df}
\alias{make.parameter.df}
\title{Make a parameter grid}
\usage{
make.parameter.df(...)
}
\arguments{
\item{...}{Named vectors of parameter values.}
}
\value{
A data.frame containing all parameter combinations.
}
\description{
Make a parameter grid
}
